refactor: Move analyzer functionality into own module
